Add files via upload
this project is meant to ameliorate the strees associated with computing some complex mathematical equation. Intend to run as a mobile app on completion, we are working to ensure it adequately serve its purpose effectively.. 

Technologies to be used are flutter (dart), Node Js, Mysql. 

We hope to launch as both Android and IOS app.